requirements.txt
jdatetime
PyYAML~=6.0.1
colorama~=0.4.6
httpx~=0.27.0
python-telegram-bot
aiomysql